Update README.md
Made some adjustments to the contents. The readme is not done. It still needs information about getting involved, contributing to the implementation and a description of current members in the collaboration. 

Further, a document describing the use of RFCs (and potentially other documents for decision-making) is missing, and must be linked to from here.