Suggestions: Update README.md (#3)
* Update README.md

Some proposals for changes. based on inputs from the internal team, and with a structure based on the [InnerSource Commons Base Documentation Pattern](https://patterns.innersourcecommons.org/p/base-documentation)

* Update README.md

Corrected the content to a more specific scope, and inserted a missing link.

* Update README.md

Made some adjustments to the contents. The readme is not done. It still needs information about getting involved, contributing to the implementation and a description of current members in the collaboration. 

Further, a document describing the use of RFCs (and potentially other documents for decision-making) is missing, and must be linked to from here.

* Update README.md

Modified the first few sentences, to include green transition goals, and open source structure setup

* Update profile/README.md

Co-authored-by: Johannes Tigges <lenucksi@users.noreply.github.com>

* Update README.md

Made some changes to the first paragraph based on workshop inputs

* Update README.md

Added some information in the "Getting Involved" section, and "Who Are We?" section, and applied some changes, as suggested by @lenucksi

* Update README.md

Added project boards for Registry and internal collaboration

* Update README.md

* Update profile/README.md

Co-authored-by: Johannes Tigges <lenucksi@users.noreply.github.com>