Update README.md
Made some small modifications, as this is the one showed on the public "Overview" page. I have created a similar one for the member "Overview" page, where information about our private repositories is included, to make information in the readme as concise and relevant as possible for the reader.

Additionally added a link to directly see issues marked with "help wanted".